requirements.txt
Flask
request
Flask-WTF
bcrypt
Flask-login
wtforms
ConfigParser
python-owasp-zap-v2.4